refactored application for better separation of concerns and readability
